Moved plot_against_time to plotting.cartesian submodule
